fix(comm): change timeout constants to use uint16_t type

+++ src/libcommon/comm.h
@@ -29,9 +29,9 @@ inline constexpr char localHostName[] = "kDrive-localhost";
 inline constexpr char serverCertSubject[] = "kdrive-server-localhost";
 inline constexpr char clientCertSubject[] = "kdrive-client-localhost";
 
-inline constexpr uint8_t COMM_SHORT_TIMEOUT = 1000;
-inline constexpr uint8_t COMM_AVERAGE_TIMEOUT = 10000;
-inline constexpr uint8_t COMM_LONG_TIMEOUT = 60000;
+inline constexpr uint16_t COMM_SHORT_TIMEOUT = 1000;
+inline constexpr uint16_t COMM_AVERAGE_TIMEOUT = 10000;
+inline constexpr uint16_t COMM_LONG_TIMEOUT = 60000;
 
 inline constexpr char MSG_TYPE[] = "type";
